Completed up to section 9.2.
At this point, almost all the sections up to section 9.2 (programs in
command line shells) has been completed. The only things that have not
been completed are a few method definitions in the Java section, but
this will be fixed in the editing phase (I just don't want to get into
writing those method descriptions right now). I have also completed the
resources section, which contains some interesting information that was
not worth of being included in the main chapters. Given the current
completion of the book, in my estimation a rough first draft should be
completed within a month depending on how busy school is and if I end up
messing around. At that point I will probably submit it to my teacher as
my side project for the year and gather feedback from proofing. The
majority of the book, as it is now 55 pages long, is already completed.
I'm guessing the completed book will be ~70 pages depending on how long
the rest of the shell script section and the C section will take.
Labelling each paragraph with a summary throughout the book in its
early stages was a really good idea, and I will use it for reports I
write in the future. The editing phase will be a behemouth, as there are
many places where there are gramatically incorrect sentences and messy
formatting, but I should worry about that only after I have fully
completed the first draft.